requirements.txt
Flask==0.10.1
Flask-HTTPAuth==2.3.0
Flask-RESTful==0.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-SSLify==0.1.5
itsdangerous==0.24
Jinja2==2.7.3
python-ldap==2.4.10
MarkupSafe==0.23
pi==0.1.2
pyasn1==0.1.7
cffi==1.0.3
pyopenssl==0.15.1
pycurl==7.19.5.1
pytz==2015.2
six==1.9.0
SQLAlchemy==0.9.7
Werkzeug==0.9.6
shell==1.0.1
